+++ scripts/UniKP_kcat.py
@@ -56,16 +56,20 @@ if __name__ == '__main__':
         else:
             Label[i] = math.log(Label[i], 10)
     print(max(Label), min(Label))
+    
     # Feature Extractor
-    smiles_input = smiles_to_vec(Smiles, device=device)
-    sequence_input = Seq_to_vec(sequence, device=device)
-    feature = np.concatenate((smiles_input, sequence_input), axis=1)
-
-    model_path=os.path.join(script_path,'..','retrained','features_16838_PreKcat.pkl')
-    os.makedirs(os.path.dirname(model_path),exist_ok=True)
-    with open(model_path, "wb") as f:
-
-        pickle.dump(feature, f)
+    feature_path=os.path.join(script_path,'..','retrained','features_16838_PreKcat.pkl')
+    if os.path.exists(feature_path):
+        smiles_input = smiles_to_vec(Smiles, device=device)
+        sequence_input = Seq_to_vec(sequence, device=device)
+        feature = np.concatenate((smiles_input, sequence_input), axis=1)
+        os.makedirs(os.path.dirname(feature_path),exist_ok=True)
+        with open(feature_path, "wb") as f:
+            pickle.dump(feature, f)
+    else:
+        with open(feature_path, 'rb') as f:
+            feature=pickle.load(f)
+   
     Label = np.array(Label)
     # Input dataset
     feature_new = []
